Feature/update thumbs on change (#14)
Thumbnails are now automatically created when media is added to the media directory.

This is the default behavior. Can be disabled with the new configuration parameter "genthumbsonadd".

Also, on the WEB UI there will be a warning sign when thumbnails are generated (shown either on startup or when files are added to the media directory).